+++ packages/contracts/contracts/voting/admin/AdminSetup.sol
@@ -19,8 +19,8 @@ contract AdminSetup is PluginSetup {
     /// @notice The address of `Admin` plugin logic contract to be cloned.
     address private immutable implementation;
 
-    /// @notice The address zero to be used as oracle address for permissions.
-    address private constant NO_ORACLE = address(0);
+    /// @notice The address zero to be used as condition address for permissions.
+    address private constant NO_CONDITION = address(0);
 
     /// @notice Thrown if admin address is zero.
     /// @param admin The admin address.
@@ -71,7 +71,7 @@ contract AdminSetup is PluginSetup {
             PermissionLib.Operation.Grant,
             plugin,
             admin,
-            NO_ORACLE,
+            NO_CONDITION,
             Admin(plugin).EXECUTE_PROPOSAL_PERMISSION_ID()
         );
 
@@ -80,7 +80,7 @@ contract AdminSetup is PluginSetup {
             PermissionLib.Operation.Grant,
             _dao,
             plugin,
-            NO_ORACLE,
+            NO_CONDITION,
             DAO(payable(_dao)).EXECUTE_PERMISSION_ID()
         );
     }
@@ -108,7 +108,7 @@ contract AdminSetup is PluginSetup {
             PermissionLib.Operation.Revoke,
             _dao,
             _plugin,
-            NO_ORACLE,
+            NO_CONDITION,
             DAO(payable(_dao)).EXECUTE_PERMISSION_ID()
         );
     }